humanitiesdata/app/requirements.txt
numpy
mysql-connector-python
